# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FixYa - Plataforma de servicios técnicos para el hogar
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [usuario, tecnico, administrador]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── SERVICE CATEGORIES ──────────────────────────────────────────────────────
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all service categories
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service category by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateService
      tags: [services]
      summary: Update service category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service category (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── TECHNICIANS ─────────────────────────────────────────────────────────────
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List technicians
      parameters:
        - name: serviceId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pendiente, aprobado, rechazado, suspendido]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of technicians
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianList"
  /technicians/me:
    get:
      operationId: getMyTechnicianProfile
      tags: [technicians]
      summary: Get current technician profile
      responses:
        "200":
          description: Technician profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createTechnicianProfile
      tags: [technicians]
      summary: Register as a technician
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created technician profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
    patch:
      operationId: updateMyTechnicianProfile
      tags: [technicians]
      summary: Update current technician profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated technician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get technician by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "404":
          $ref: "#/components/responses/NotFound"
  /technicians/{id}/approve:
    post:
      operationId: approveTechnician
      tags: [technicians]
      summary: Approve technician (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}/reject:
    post:
      operationId: rejectTechnician
      tags: [technicians]
      summary: Reject technician (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Technician rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/top:
    get:
      operationId: getTopTechnicians
      tags: [technicians]
      summary: Get top-rated technicians
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 6
      responses:
        "200":
          description: Top technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"

  # ─── SERVICE REQUESTS ─────────────────────────────────────────────────────────
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── USER ──────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        telefono:
          type: ["string", "null"]
        direccion:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [usuario, tecnico, administrador]
        activo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - nombre
        - apellido
        - email
        - role
        - activo
        - createdAt
    UserList:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - users
        - total
        - page
        - limit
    UserUpdate:
      type: object
      properties:
        nombre:
          type: string
        apellido:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        avatarUrl:
          type: string
    AdminUserUpdate:
      type: object
      properties:
        role:
          type: string
          enum: [usuario, tecnico, administrador]
        activo:
          type: boolean

    # ─── SERVICE ───────────────────────────────────────────────────────────────
    Service:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: ["string", "null"]
        icono:
          type: ["string", "null"]
        precioBase:
          type: number
        activo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nombre
        - precioBase
        - activo
        - createdAt
    ServiceInput:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        icono:
          type: string
        precioBase:
          type: number
      required:
        - nombre
        - precioBase
    ServiceUpdate:
      type: object
      properties:
        nombre:
          type: string
        descripcion:
          type: string
        icono:
          type: string
        precioBase:
          type: number
        activo:
          type: boolean

    # ─── TECHNICIAN ────────────────────────────────────────────────────────────
    Technician:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        especialidad:
          type: string
        descripcion:
          type: ["string", "null"]
        experienciaAnios:
          type: integer
        precioHora:
          type: number
        disponible:
          type: boolean
        status:
          type: string
          enum: [pendiente, aprobado, rechazado, suspendido]
        promedioCalificacion:
          type: number
        totalTrabajos:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        service:
          $ref: "#/components/schemas/Service"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - serviceId
        - especialidad
        - experienciaAnios
        - precioHora
        - disponible
        - status
        - promedioCalificacion
        - totalTrabajos
        - createdAt
    TechnicianList:
      type: object
      properties:
        technicians:
          type: array
          items:
            $ref: "#/components/schemas/Technician"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - technicians
        - total
        - page
        - limit
    TechnicianInput:
      type: object
      properties:
        serviceId:
          type: integer
        especialidad:
          type: string
        descripcion:
          type: string
        experienciaAnios:
          type: integer
        precioHora:
          type: number
      required:
        - serviceId
        - especialidad
        - experienciaAnios
        - precioHora
    TechnicianUpdate:
      type: object
      properties:
        especialidad:
          type: string
        descripcion:
          type: string
        experienciaAnios:
          type: integer
        precioHora:
          type: number
        disponible:
          type: boolean
    RejectInput:
      type: object
      properties:
        motivo:
          type: string
      required:
        - motivo

    # ─── SERVICE REQUEST ───────────────────────────────────────────────────────
...